requirements.txt
pyserial
pymodbus
PyQt5
